+++ .eslintignore
@@ -1,5 +1,4 @@
 **/node_modules
 **/coverage
 packages/generator-eslint-shopify/generators
-packages/shopify-codemods/transforms
-packages/shopify-codemods/test/fixtures
+packages/shopify-codemod/test/fixtures
