+++ package.json
@@ -4,7 +4,7 @@
   "description": "The home for all things JavaScript at Shopify.",
   "scripts": {
     "lint": "eslint . --max-warnings 0",
-    "test": "cd packages/eslint-plugin-shopify && npm test && cd ../generator-eslint-shopify && npm test && cd ../shopify-codemods && npm test && cd ../..",
+    "test": "cd packages/eslint-plugin-shopify && npm test && cd ../generator-eslint-shopify && npm test && cd ../shopify-codemod && npm test && cd ../..",
     "check": "npm run lint && npm test",
     "setup": "npm install && lerna bootstrap",
     "updated": "lerna updated",
