+++ packages/shopify-codemod/.npmignore
@@ -1,2 +1,3 @@
 test
+coverage
 .eslintignore
