+++ packages/shopify-codemod/package.json
@@ -1,17 +1,14 @@
 {
-  "name": "shopify-codemods",
+  "name": "shopify-codemod",
   "version": "10.1.1",
-  "description": "",
-  "main": "index.js",
+  "description": "Custom JSCodeShift transforms to help convert our rusty JavaScript.",
+  "main": "transforms/index.js",
   "scripts": {
-    "test": "NODE_PATH=./src:./test:$NODE_PATH mocha 'test/**/*.test.js' --compilers js:babel-core/register",
+    "test": "NODE_PATH=./transforms:./test:$NODE_PATH mocha 'test/**/*.test.js' --compilers js:babel-core/register",
     "test:watch": "npm run test -- --watch --reporter dot",
-    "test:cover": "NODE_PATH=$NODE_PATH:./src:./test babel-node $(npm bin)/isparta cover --root transforms/ --reporter text --reporter html $(npm bin)/_mocha -- --reporter spec test/**/*.test.js",
-    "clean": "rm -rf transforms/ coverage/",
-    "build:lib": "babel src --out-dir transforms",
-    "build": "npm run clean && npm run build:lib",
-    "preversion": "npm test",
-    "prepublish": "npm test && npm run build"
+    "test:cover": "NODE_PATH=$NODE_PATH:./transforms:./test babel-node $(npm bin)/isparta cover --root transforms/ --reporter text --reporter html $(npm bin)/_mocha -- --reporter spec test/**/*.test.js",
+    "clean": "rm -rf coverage/",
+    "prepublish": "npm test"
   },
   "babel": {
     "presets": [
@@ -24,7 +21,6 @@
   "author": "Chris Sauve <chrismsauve@gmail.com>",
   "license": "ISC",
   "devDependencies": {
-    "babel-cli": "^6.6.5",
     "babel-core": "^6.7.2",
     "babel-preset-shopify": "file:../babel-preset-shopify",
     "chai": "^3.5.0",
@@ -33,6 +29,5 @@
     "mocha": "^2.4.5"
   },
   "dependencies": {
-    "lodash": "^4.6.1"
   }
 }
