+++ packages/shopify-codemod/transforms/index.js
@@ -1,0 +1,1 @@
+export {default as mochaContextToClosure} from './mocha-context-to-closure';
